refactor: 공용 폴더 구조 변경
