+++ src/components/Auth/Copyright/Copyright.module.scss
