+++ src/components/Auth/Copyright/index.tsx
